Add support for 65th sql flag

+++ WowPacketParser/Store/StoreDictionaries.cs
@@ -12,7 +12,7 @@ namespace WowPacketParser.Store
 {
     public abstract class Store
     {
-        public static ulong SQLEnabledFlags { protected get; set; }
+        public static UInt128 SQLEnabledFlags { protected get; set; }
         public List<SQLOutput> Types { get; protected set; }
 
         protected bool ProcessFlags()
